#64 checkbox redux
